Added TOC generating in documentations

+++ tools/gendoc.py
@@ -1,13 +1,48 @@
-from distutils import extension
-from pydoc import doc
+# metapp library
+
+# Copyright (C) 2022 Wang Qi (wqking)
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#   http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 from cpp2md import extractMarkdownFromCpp
 import os
+import sys
 import glob
 import re
 import shutil
+import subprocess
+import shlex
 
 sourcePath = '../tests/docsrc'
 docPath = '../doc'
+
+def isFileUpToDate(sourceFile, targetFile) :
+	if not os.path.exists(targetFile) :
+		return False
+	if os.path.getmtime(sourceFile) > os.path.getmtime(targetFile) :
+		return False
+	return True
+
+def isWindows() :
+	return sys.platform.startswith('win')
+
+def normalizeCommand(command) :
+	return shlex.split(command, posix = not isWindows())
+
+def doPostProcessMarkdown(file) :
+	command = 'perl addtoc2md.pl --max-level=4 "%s"' % (file)
+	command = normalizeCommand(command)
+	subprocess.run(command)
 
 def doProcessFile(sourceFile) :
 	matches = re.search(r'doc_([^\\\/]+)\.(\w+)$', sourceFile)
@@ -22,13 +57,19 @@ def doProcessFile(sourceFile) :
 	targetPath = os.path.join(docPath, relativePath)
 	targetFileName = os.path.join(targetPath, fileName + '.md')
 	os.makedirs(targetPath, exist_ok = True)
-	print(sourceFile)
+	if isFileUpToDate(sourceFile, targetFileName) :
+		print("%s is up to date, skip." % (targetFileName))
+		return
+	print("Generate %s" % (targetFileName))
 	if extension == 'cpp' :
 		extractMarkdownFromCpp(sourceFile, targetFileName)
 	elif extension in [ 'md' ] :
 		shutil.copy(sourceFile, targetFileName)
 	else :
 		print("Unknow file type: %s" % (extension))
+		return
+	if targetFileName.endswith('.md') :
+		doPostProcessMarkdown(targetFileName)
 
 def doMain() :
 	global sourcePath, docPath
